# src/main/resources/requests/Historia-Medica_Swagger.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/updateEventos/{eventoId}:
    put:
      tags:
      - google-calendar-controller-rest
      operationId: updateGoogleMeetingHandler
      parameters:
      - name: eventoId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleCalendarioPostRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/addEventos:
    post:
      tags:
      - google-calendar-controller-rest
      operationId: createGoogleMeetingHandler
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleCalendarioPostRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/login/google:
    get:
      tags:
      - google-calendar-controller-rest
      operationId: googleConnectionStatusHandler_1
      parameters:
      - name: code
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                oneOf:
                - type: object
                - $ref: '#/components/schemas/RedirectView'
...
  /api/getListEventos:
    get:
      tags:
      - google-calendar-controller-rest
      operationId: getListCalendarEventsHandler
      parameters:
      - name: datosListEventos
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/GoogleCalendarioGetListRequest'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
  /api/getEventos/{eventoId}:
    get:
      tags:
      - google-calendar-controller-rest
      operationId: getCalendarEventHandler
      parameters:
      - name: eventoId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/deleteEventos/{eventoId}:
    delete:
      tags:
      - google-calendar-controller-rest
      operationId: deleteGoogleMeetingHandler
      parameters:
      - name: eventoId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                properties:
                  classInfo:
                    $ref: '#/components/schemas/ClassInfo'
                  anyoneCanAddSelf:
                    type: boolean
                  attachments:
                    type: array
                    items:
                      $ref: '#/components/schemas/EventAttachment'
                  attendees:
                    type: array
                    items:
                      $ref: '#/components/schemas/EventAttendee'
                  attendeesOmitted:
                    type: boolean
                  colorId:
                    type: string
                  conferenceData:
                    type: object
                    properties:
                      classInfo:
                        $ref: '#/components/schemas/ClassInfo'
                      conferenceId:
                        type: string
                      conferenceSolution:
                        type: object
                        properties:
                          classInfo:
                            $ref: '#/components/schemas/ClassInfo'
                          iconUri:
                            type: string
                          key:
                            type: object
                            properties:
                              classInfo:
                                $ref: '#/components/schemas/ClassInfo'
                              type:
                                type: string
                              factory:
                                $ref: '#/components/schemas/JsonFactory'
                              unknownKeys:
                                type: object
                                additionalProperties:
                                  type: object
                              empty:
                                type: boolean
                            additionalProperties:
                              type: object
                          name:
                            type: string
                          factory:
                            $ref: '#/components/schemas/JsonFactory'
                          unknownKeys:
                            type: object
                            additionalProperties:
                              type: object
                          empty:
                            type: boolean
                        additionalProperties:
                          type: object
                      createRequest:
                        type: object
                        properties:
                          classInfo:
                            $ref: '#/components/schemas/ClassInfo'
                          conferenceSolutionKey:
                            type: object
                            properties:
                              classInfo:
                                $ref: '#/components/schemas/ClassInfo'
                              type:
                                type: string
                              factory:
                                $ref: '#/components/schemas/JsonFactory'
                              unknownKeys:
                                type: object
                                additionalProperties:
                                  type: object
                              empty:
                                type: boolean
                            additionalProperties:
                              type: object
                          requestId:
                            type: string
                          status:
                            type: object
                            properties:
                              classInfo:
                                $ref: '#/components/schemas/ClassInfo'
                              statusCode:
                                type: string
                              factory:
                                $ref: '#/components/schemas/JsonFactory'
                              unknownKeys:
                                type: object
                                additionalProperties:
                                  type: object
                              empty:
                                type: boolean
                            additionalProperties:
                              type: object
                          factory:
                            $ref: '#/components/schemas/JsonFactory'
                          unknownKeys:
                            type: object
                            additionalProperties:
                              type: object
                          empty:
                            type: boolean
                        additionalProperties:
                          type: object
                      entryPoints:
                        type: array
                        items:
                          $ref: '#/components/schemas/EntryPoint'
                      notes:
                        type: string
                      parameters:
                        type: object
                        properties:
                          classInfo:
                            $ref: '#/components/schemas/ClassInfo'
                          addOnParameters:
                            type: object
                            properties:
                              classInfo:
                                $ref: '#/components/schemas/ClassInfo'
                              parameters:
                                type: object
                                additionalProperties:
                                  type: string
                              factory:
                                $ref: '#/components/schemas/JsonFactory'
                              unknownKeys:
                                type: object
                                additionalProperties:
                                  type: object
                              empty:
                                type: boolean
                            additionalProperties:
                              type: object
                          factory:
                            $ref: '#/components/schemas/JsonFactory'
                          unknownKeys:
                            type: object
                            additionalProperties:
                              type: object
                          empty:
                            type: boolean
                        additionalProperties:
                          type: object
                      signature:
                        type: string
                      factory:
                        $ref: '#/components/schemas/JsonFactory'
                      unknownKeys:
                        type: object
                        additionalProperties:
                          type: object
                      empty:
                        type: boolean
                    additionalProperties:
                      type: object
                  created:
                    $ref: '#/components/schemas/DateTime'
                  creator:
                    type: object
                    properties:
                      classInfo:
                        $ref: '#/components/schemas/ClassInfo'
                      displayName:
                        type: string
                      email:
                        type: string
                      id:
                        type: string
                      self:
                        type: boolean
                      factory:
                        $ref: '#/components/schemas/JsonFactory'
                      unknownKeys:
                        type: object
                        additionalProperties:
                          type: object
                      empty:
                        type: boolean
                    additionalProperties:
                      type: object
                  description:
                    type: string
                  end:
                    type: object
                    properties:
                      classInfo:
                        $ref: '#/components/schemas/ClassInfo'
                      date:
                        $ref: '#/components/schemas/DateTime'
                      dateTime:
                        $ref: '#/components/schemas/DateTime'
                      timeZone:
                        type: string
                      factory:
                        $ref: '#/components/schemas/JsonFactory'
                      unknownKeys:
                        type: object
                        additionalProperties:
                          type: object
                      empty:
                        type: boolean
                    additionalProperties:
                      type: object
                  endTimeUnspecified:
                    type: boolean
                  etag:
                    type: string
                  eventType:
                    type: string
                  extendedProperties:
                    type: object
                    properties:
                      classInfo:
                        $ref: '#/components/schemas/ClassInfo'
                      shared:
                        type: object
                        additionalProperties:
                          type: string
                      private:
                        type: object
                        additionalProperties:
                          type: string
                      factory:
                        $ref: '#/components/schemas/JsonFactory'
                      unknownKeys:
                        type: object
                        additionalProperties:
                          type: object
                      empty:
                        type: boolean
                    additionalProperties:
                      type: object
                  gadget:
                    type: object
                    properties:
                      classInfo:
                        $ref: '#/components/schemas/ClassInfo'
                      display:
                        type: string
                      height:
                        type: integer
                        format: int32
                      iconLink:
                        type: string
                      link:
                        type: string
                      preferences:
                        type: object
                        additionalProperties:
                          type: string
                      title:
                        type: string
                      type:
                        type: string
                      width:
                        type: integer
                        format: int32
                      factory:
                        $ref: '#/components/schemas/JsonFactory'
                      unknownKeys:
                        type: object
                        additionalProperties:
                          type: object
                      empty:
                        type: boolean
                    additionalProperties:
                      type: object
                  guestsCanInviteOthers:
                    type: boolean
                  guestsCanModify:
                    type: boolean
                  guestsCanSeeOtherGuests:
                    type: boolean
                  hangoutLink:
                    type: string
                  htmlLink:
                    type: string
                  id:
                    type: string
                  kind:
                    type: string
                  location:
                    type: string
                  locked:
                    type: boolean
                  organizer:
                    type: object
                    properties:
                      classInfo:
                        $ref: '#/components/schemas/ClassInfo'
                      displayName:
                        type: string
                      email:
                        type: string
                      id:
                        type: string
                      self:
                        type: boolean
                      factory:
                        $ref: '#/components/schemas/JsonFactory'
                      unknownKeys:
                        type: object
                        additionalProperties:
                          type: object
                      empty:
                        type: boolean
                    additionalProperties:
                      type: object
                  originalStartTime:
                    type: object
                    properties:
                      classInfo:
                        $ref: '#/components/schemas/ClassInfo'
                      date:
                        $ref: '#/components/schemas/DateTime'
                      dateTime:
                        $ref: '#/components/schemas/DateTime'
                      timeZone:
                        type: string
                      factory:
                        $ref: '#/components/schemas/JsonFactory'
                      unknownKeys:
                        type: object
                        additionalProperties:
                          type: object
                      empty:
                        type: boolean
                    additionalProperties:
                      type: object
                  privateCopy:
                    type: boolean
                  recurrence:
                    type: array
                    items:
                      type: string
                  recurringEventId:
                    type: string
                  reminders:
                    type: object
                    properties:
                      classInfo:
                        $ref: '#/components/schemas/ClassInfo'
                      overrides:
                        type: array
                        items:
                          $ref: '#/components/schemas/EventReminder'
                      useDefault:
                        type: boolean
                      factory:
                        $ref: '#/components/schemas/JsonFactory'
                      unknownKeys:
                        type: object
                        additionalProperties:
                          type: object
                      empty:
                        type: boolean
                    additionalProperties:
                      type: object
                  sequence:
                    type: integer
                    format: int32
                  source:
                    type: object
                    properties:
                      classInfo:
                        $ref: '#/components/schemas/ClassInfo'
                      title:
                        type: string
                      url:
                        type: string
                      factory:
                        $ref: '#/components/schemas/JsonFactory'
                      unknownKeys:
                        type: object
                        additionalProperties:
                          type: object
                      empty:
                        type: boolean
                    additionalProperties:
                      type: object
                  start:
                    type: object
                    properties:
                      classInfo:
                        $ref: '#/components/schemas/ClassInfo'
                      date:
                        $ref: '#/components/schemas/DateTime'
                      dateTime:
                        $ref: '#/components/schemas/DateTime'
                      timeZone:
                        type: string
                      factory:
                        $ref: '#/components/schemas/JsonFactory'
                      unknownKeys:
                        type: object
                        additionalProperties:
                          type: object
                      empty:
                        type: boolean
                    additionalProperties:
                      type: object
                  status:
                    type: string
                  summary:
                    type: string
                  transparency:
                    type: string
                  updated:
                    $ref: '#/components/schemas/DateTime'
                  visibility:
                    type: string
                  icalUID:
                    type: string
                  factory:
                    $ref: '#/components/schemas/JsonFactory'
                  unknownKeys:
                    type: object
                    additionalProperties:
                      type: object
                  empty:
                    type: boolean
                additionalProperties:
                  type: object
components:
  schemas:
...
    GoogleCalendarioPostRequest:
      type: object
      properties:
        summary:
          type: string
        location:
          type: string
        description:
          type: string
        attendees:
          type: array
          items:
            type: string
        startTime:
          type: string
        endTime:
          type: string
...
    RedirectView:
      type: object
      properties:
        applicationContext:
          $ref: '#/components/schemas/ApplicationContext'
        servletContext:
          type: object
          properties:
            classLoader:
              type: object
              properties:
                parent:
                  type: object
                  properties:
                    defaultAssertionStatus:
                      type: boolean
                      writeOnly: true
                defaultAssertionStatus:
                  type: boolean
                  writeOnly: true
            majorVersion:
              type: integer
              format: int32
            initParameterNames:
              type: object
            effectiveMajorVersion:
              type: integer
              format: int32
            attributeNames:
              type: object
            servletContextName:
              type: string
            effectiveMinorVersion:
              type: integer
              format: int32
            jspConfigDescriptor:
              type: object
              properties:
                jspPropertyGroups:
                  type: array
                  items:
                    type: object
                    properties:
                      buffer:
                        type: string
                      defaultContentType:
                        type: string
                      urlPatterns:
                        type: array
                        items:
                          type: string
                      scriptingInvalid:
                        type: string
                      errorOnUndeclaredNamespace:
                        type: string
                      includePreludes:
                        type: array
                        items:
                          type: string
                      trimDirectiveWhitespaces:
                        type: string
                      deferredSyntaxAllowedAsLiteral:
                        type: string
                      includeCodas:
                        type: array
                        items:
                          type: string
                      pageEncoding:
                        type: string
                      isXml:
                        type: string
                      elIgnored:
                        type: string
                taglibs:
                  type: array
                  items:
                    type: object
                    properties:
                      taglibLocation:
                        type: string
                      taglibURI:
                        type: string
            sessionCookieConfig:
              type: object
              properties:
                name:
                  type: string
                path:
                  type: string
                domain:
                  type: string
                comment:
                  type: string
                secure:
                  type: boolean
                httpOnly:
                  type: boolean
                maxAge:
                  type: integer
                  format: int32
            effectiveSessionTrackingModes:
              uniqueItems: true
              type: array
              items:
                type: string
                enum:
                - COOKIE
                - URL
                - SSL
            virtualServerName:
              type: string
            responseCharacterEncoding:
              type: string
            sessionTimeout:
              type: integer
              format: int32
            requestCharacterEncoding:
              type: string
            filterRegistrations:
              type: object
              additionalProperties:
                type: object
                properties:
                  urlPatternMappings:
                    type: array
                    items:
                      type: string
                  servletNameMappings:
                    type: array
                    items:
                      type: string
                  name:
                    type: string
                  className:
                    type: string
                  initParameters:
                    type: object
                    additionalProperties:
                      type: string
            sessionTrackingModes:
              uniqueItems: true
              type: array
              writeOnly: true
              items:
                type: string
                enum:
                - COOKIE
                - URL
                - SSL
            servletRegistrations:
              type: object
              additionalProperties:
                type: object
                properties:
                  mappings:
                    type: array
                    items:
                      type: string
                  runAsRole:
                    type: string
                  name:
                    type: string
                  className:
                    type: string
                  initParameters:
                    type: object
                    additionalProperties:
                      type: string
            defaultSessionTrackingModes:
              uniqueItems: true
              type: array
              items:
                type: string
                enum:
                - COOKIE
                - URL
                - SSL
            contextPath:
              type: string
            serverInfo:
              type: string
            minorVersion:
              type: integer
              format: int32
            servletNames:
              type: object
              deprecated: true
            servlets:
              type: object
              deprecated: true
          writeOnly: true
        contentType:
          type: string
        requestContextAttribute:
          type: string
        staticAttributes:
          type: object
          additionalProperties:
            type: object
        exposePathVariables:
          type: boolean
        exposeContextBeansAsAttributes:
          type: boolean
          writeOnly: true
        exposedContextBeanNames:
          type: array
          writeOnly: true
          items:
            type: string
        beanName:
          type: string
        url:
          type: string
        contextRelative:
          type: boolean
          writeOnly: true
        http10Compatible:
          type: boolean
          writeOnly: true
        exposeModelAttributes:
          type: boolean
          writeOnly: true
        encodingScheme:
          type: string
          writeOnly: true
        statusCode:
          type: string
          writeOnly: true
          enum:
          - 100 CONTINUE
          - 101 SWITCHING_PROTOCOLS
          - 102 PROCESSING
          - 103 CHECKPOINT
          - 200 OK
          - 201 CREATED
          - 202 ACCEPTED
          - 203 NON_AUTHORITATIVE_INFORMATION
          - 204 NO_CONTENT
          - 205 RESET_CONTENT
          - 206 PARTIAL_CONTENT
          - 207 MULTI_STATUS
          - 208 ALREADY_REPORTED
          - 226 IM_USED
          - 300 MULTIPLE_CHOICES
          - 301 MOVED_PERMANENTLY
          - 302 FOUND
          - 302 MOVED_TEMPORARILY
          - 303 SEE_OTHER
          - 304 NOT_MODIFIED
          - 305 USE_PROXY
          - 307 TEMPORARY_REDIRECT
          - 308 PERMANENT_REDIRECT
          - 400 BAD_REQUEST
          - 401 UNAUTHORIZED
          - 402 PAYMENT_REQUIRED
          - 403 FORBIDDEN
          - 404 NOT_FOUND
          - 405 METHOD_NOT_ALLOWED
          - 406 NOT_ACCEPTABLE
          - 407 PROXY_AUTHENTICATION_REQUIRED
          - 408 REQUEST_TIMEOUT
          - 409 CONFLICT
          - 410 GONE
          - 411 LENGTH_REQUIRED
          - 412 PRECONDITION_FAILED
          - 413 PAYLOAD_TOO_LARGE
          - 413 REQUEST_ENTITY_TOO_LARGE
          - 414 URI_TOO_LONG
          - 414 REQUEST_URI_TOO_LONG
          - 415 UNSUPPORTED_MEDIA_TYPE
          - 416 REQUESTED_RANGE_NOT_SATISFIABLE
          - 417 EXPECTATION_FAILED
          - 418 I_AM_A_TEAPOT
          - 419 INSUFFICIENT_SPACE_ON_RESOURCE
          - 420 METHOD_FAILURE
          - 421 DESTINATION_LOCKED
          - 422 UNPROCESSABLE_ENTITY
          - 423 LOCKED
          - 424 FAILED_DEPENDENCY
          - 425 TOO_EARLY
          - 426 UPGRADE_REQUIRED
          - 428 PRECONDITION_REQUIRED
          - 429 TOO_MANY_REQUESTS
          - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
          - 451 UNAVAILABLE_FOR_LEGAL_REASONS
          - 500 INTERNAL_SERVER_ERROR
          - 501 NOT_IMPLEMENTED
          - 502 BAD_GATEWAY
          - 503 SERVICE_UNAVAILABLE
          - 504 GATEWAY_TIMEOUT
          - 505 HTTP_VERSION_NOT_SUPPORTED
          - 506 VARIANT_ALSO_NEGOTIATES
          - 507 INSUFFICIENT_STORAGE
          - 508 LOOP_DETECTED
          - 509 BANDWIDTH_LIMIT_EXCEEDED
          - 510 NOT_EXTENDED
          - 511 NETWORK_AUTHENTICATION_REQUIRED
        expandUriTemplateVariables:
          type: boolean
          writeOnly: true
        propagateQueryParams:
          type: boolean
          writeOnly: true
        hosts:
          type: array
          items:
            type: string
        redirectView:
          type: boolean
        propagateQueryProperties:
          type: boolean
        attributes:
          type: object
          additionalProperties:
            type: string
          writeOnly: true
        attributesMap:
          type: object
          additionalProperties:
            type: object
        attributesCSV:
          type: string
          writeOnly: true
...
    GoogleCalendarioGetListRequest:
      type: object
      properties:
        startDate:
          type: string
        endDate:
          type: string
...
    JsonFactory:
      type: object
...